+++ CleverPush/CleverPush/Source/CleverPush.m
@@ -68,7 +68,7 @@
 
 @implementation CleverPush
 
-NSString * const CLEVERPUSH_SDK_VERSION = @"1.7.1";
+NSString * const CLEVERPUSH_SDK_VERSION = @"1.8.0";
 
 static BOOL registeredWithApple = NO;
 static BOOL startFromNotification = NO;