+++ CleverPush/Framework/CleverPush.framework/Versions/A/Headers/CleverPush.h
@@ -141,6 +141,9 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 + (void)setApiEndpoint:(NSString*)apiEndpoint;
 + (NSString*)getApiEndpoint;
 + (void)updateBadge:(UNMutableNotificationContent*)replacementContent;
++ (BOOL)fontFamilyExits:(NSString*)fontFamily;
++ (void)disableAppBanners;
++ (void)enableAppBanners;
 
 @end